requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
pillow==10.2.0
promise==2.3
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.9.0
